consumer2/requirements.txt
kafka-python==1.4.7
asgiref==3.2.3
Django==3.0.2
pytz==2019.3
sqlparse==0.3.0